+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
+
+### Unreleased
+
+Features
+
+* Add cop Platanus/SingularSerializer
 
 ### v0.2.0